dvgo/requirements.txt
numpy
scipy
tqdm
lpips
imageio
imageio-ffmpeg
opencv-python
torch_efficient_distloss
matplotlib